Update P11.py
P11 is a grading software that reads data such as student name, and number grade off of a text file. It then writes an new text file for the letter grades for the students. It creates an error file for any data that is not valid. It also counts how many of each letter grade was awarded.