Update Final.py
the final was an assignment in which we were given a working game. The game is simple. it picks a number between 1 and 200, but doesn't tell you. you then guess a number, and the program tells you if it is too high, or too low. this goes on until you guess the number. We were tasked with adding many new features to it. Some of these include being able to pick how many guesses you get before you lose, the ability to lose, printing how many guesses the player has used before they guess again, etc.